refactor: separate header component from dashboard detail page
